pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MiniSWERouterBench"
version = "0.1.0"
description = "MiniSWERouterBench: SWERouterBench's per-step router evaluation wired onto the mini-swe-agent scaffold."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "MiniSWERouterBench contributors" }]
keywords = ["llm", "router", "benchmark", "swe-bench", "agent", "mini-swe-agent"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "SWERouterBench>=0.2.0",
  "CommonRouterBench>=0.1.0",
  "mini-swe-agent>=2.2,<3",
  "swebench>=2.1.0",
  "docker>=7.0.0",
  "litellm>=1.50.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
miniswerouterbench = "miniswerouter.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
where = ["."]
include = ["miniswerouter*"]

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]
